improve, rewrite, refine..

+++ app/controllers/ESECtlUser.java
@@ -6,67 +6,74 @@ import play.data.validation.*;
 import models.*;
 
 @With(Secure.class)
-public class usr extends Controller
+public class ESECtlUser extends Controller
 {
-	public static void ls_cals (
-		String id
+	private static String auth_user = Secure.Security.connected();
+
+	public static void lsCalendars (
+		String uri_user
 	) {
 		ESEUser u;
 		List<ESECalendar> lc;
-		String authid = Secure.Security.connected();
-		String user = id==null ?authid :id;
+		String user = uri_user==null ?auth_user :uri_user;
 
 		if ((u = ESEUser.getUser(user)) == null) {
-			user = authid;
+			user = auth_user;
 			u = ESEUser.getUser(user);
 		}
 		lc = u.getAllCalendars();
-		render(lc);
+		render(user, lc);
 	}
 
-	public static void ls (
+	public static void lsUsers (
 	) {
 		List<ESEUser> lu;
-		String authid = Secure.Security.connected();
-
-		lu = ESEUser.getAllOtherUsers(authid);
+		lu = ESEUser.getAllOtherUsers(auth_user);
 		render(lu);
 	}
 
-	public static void add_usr (
-		@Required String uname
+	public static void addUser (
+		String uname
 	) {
 		render(uname);
 	}
 
-	public static void add_usr_post (
+	public static void addUserPost (
 		@Required String uname,
-		@Required String pass,
-		String fname,
-		String lname
+		@Required String upass,
+		String ufname,
+		String ulname
 	) {
 		ESEUser u = null;
 		if (!validation.hasErrors()) {
 			if ((u = ESEUser.getUser(uname)) != null) {
-				u.editPassword(pass);
-				u.editFirstName(pass);
-				u.editFamilyName(pass);
+				u.editPassword(upass);
+				u.editFirstName(upass);
+				u.editFamilyName(upass);
 			}
 			else {
-				ESEFactory.createUser(uname, pass, fname, lname);
+				ESEFactory.createUser(uname, upass,
+					ufname, ufname);
 			}
-			usr.ls();
+			ESECtlUser.lsUsers();
 		}
 		params.flash();
 		validation.keep();
-		add_usr(uname);
+		ESECtlUser.addUser(uname);
 	}
 
-	public static void rm_usr (
+	public static void delUser (
 		String uname
 	) {
-		ESEUser u = ESEUser.find("byUsername", uname).first();
-		u.delete();
-		usr.ls();
+		ESEUser u = ESEUser.getUser(uname);
+		if (u != null) {
+			/**
+			 *	XXX: ESEUser.removeUser() should exist
+			 *	and not only delete the user but also
+			 *	all his/her Calendars and/or Events..
+			 */
+			u.delete();
+		}
+		ESECtlUser.lsUsers();
 	}
 }
